test: split app.e2e by surface, one fresh-session helper, a declared mock scenario table
- app.e2e.ts (3,813 lines, eight suites on one mutable tab) loses its
  thirteen already-hermetic tests to files named by surface —
  document.e2e.ts (the live response document + subagent decks),
  activity.e2e.ts (indicator + follow-the-tail), shell-effects.e2e.ts
  (completions, needs-you toasts, file drop, the mermaid sandbox, the
  badged notice) — each owning a fresh daemon and page. The remaining
  shared-page tests stay together (2,283 lines).
- e2e-harness.ts now owns withFreshMockSession (two suites had their own
  with different signatures), typePrompt, PHONE_CONTEXT, waitTurnIdle and
  settled; ui-contracts uses it too.
- itest-harness: startDaemon rejects the moment the child exits before
  listening (with its log) instead of a generic 15 s timeout;
  createSession takes { cwd, token } so the three hand-rolled wire-seeded
  session creations in app.e2e and changes.e2e are one helper.
- Three positive-sync sleeps (a 300 ms nap in the Changes file selector
  called ~15×, two "let the fan-out finish" naps) wait on the thing they
  were waiting for.
- mock.ts: the 28-line order-sensitive regex chain in pushPrompt is a
  declared MockSession.SCENARIOS table (id, canonical prompt, matcher,
  play) with MockSession.prompts / scenarioFor for tests; a Tier-1 test
  proves every canonical prompt routes to its own scenario and that plain
  prose falls through to the template deck — the dispatch-order guard the
  suite lacked.
- Direct unit tests for AsyncQueue and shouldFocusPromptFromTranscriptPointer.

Unit 914/914, Tier-2 152/152.
