+++ server/testing/app.e2e.ts
@@ -5,8 +5,8 @@ import { createServer } from "node:http";
 import os from "node:os";
 import path from "node:path";
 import { type Browser, type Locator, type Page } from "playwright-core";
-import { fixtureGit as git, startDaemon, TestClient, type Daemon } from "./itest-harness";
-import { assertAxeClean, launchChrome, noSideScroll } from "./e2e-harness";
+import { createSession, fixtureGit as git, startDaemon, TestClient, type Daemon } from "./itest-harness";
+import { assertAxeClean, launchChrome, noSideScroll, typePrompt, withFreshMockSession } from "./e2e-harness";
 import type { ClientMsg } from "../protocol";
 import { startOllamaFixture } from "./ollama-fixture";
 import { startRelayStub } from "../relay/relay-stub";
@@ -137,170 +137,7 @@ after(async () => {
   await d?.stop();
 });
 
-async function withFreshMockSession(
-  token: string,
-  run: (isolatedPage: Page) => Promise<void>,
-  agent = "Claude Code",
-  // Runs before the first navigation — the slot for addInitScript stubs that
-  // must exist before the app boots (the NF Notification recorder).
-  prepare?: (isolatedPage: Page) => Promise<void>,
-): Promise<void> {
-  const daemon = await startDaemon({ MIRAFOLD_TOKEN: token });
-  let isolatedPage: Page | undefined;
-  try {
-    isolatedPage = await browser.newPage();
-    if (prepare) await prepare(isolatedPage);
-    await isolatedPage.goto(`http://127.0.0.1:${daemon.port}/?token=${token}`);
-    await isolatedPage.locator(".onb-agent", { hasText: agent }).click();
-    await isolatedPage.waitForURL(/\/s\/[\w-]+/);
-    await run(isolatedPage);
-  } finally {
-    try {
-      await isolatedPage?.close();
-    } finally {
-      await daemon.stop();
-    }
-  }
-}
 
-async function submitPrompt(p: Page, text: string): Promise<Locator> {
-  const prompt = p.locator(".prompt-box textarea");
-  await prompt.fill(text);
-  await prompt.press("Enter");
-  return prompt;
-}
-
-async function settleLiveDocument(p: Page): Promise<Locator> {
-  const prompt = await submitPrompt(p, "live document demo");
-  await p
-    .locator(".turn-assistant", { hasText: "response finished as one live composition" })
-    .waitFor({ timeout: 30_000 });
-  await p.locator(".activity-line").waitFor({ state: "detached", timeout: 15_000 });
-  return prompt;
-}
-
-function readLiveDocumentPresentation(p: Page) {
-  return p.evaluate(() => {
-    const userTurn = document.querySelector(".turn-user") as HTMLElement | null;
-    const response = document.querySelector(".response-document") as HTMLElement | null;
-    const prose = response?.querySelector(".turn-assistant") as HTMLElement | null;
-    const card = response?.querySelector(".rc-card") as HTMLElement | null;
-    const wideTable = document.querySelector(".response-document .rc-table") as HTMLElement | null;
-    const h1 = document.querySelector(".response-document h1") as HTMLElement | null;
-    const h2 = document.querySelector(".response-document h2") as HTMLElement | null;
-    const h3 = document.querySelector(".response-document h3") as HTMLElement | null;
-    const quote = document.querySelector(".response-document blockquote") as HTMLElement | null;
-    const code = document.querySelector(
-      ".response-document .markdown pre code.hljs",
-    ) as HTMLElement | null;
-    const markdownTable = document.querySelector(
-      ".response-document .markdown-table-scroll",
-    ) as HTMLElement | null;
-    if (
-      !userTurn ||
-      !response ||
-      !prose ||
-      !card ||
-      !wideTable ||
-      !h1 ||
-      !h2 ||
-      !h3 ||
-      !quote ||
-      !code ||
-      !markdownTable
-    ) {
-      return null;
-    }
-    const userTurnRect = userTurn.getBoundingClientRect();
-    const responseRect = response.getBoundingClientRect();
-    const proseRect = prose.getBoundingClientRect();
-    const h1Style = getComputedStyle(h1);
-    const h2Style = getComputedStyle(h2);
-    const h3Style = getComputedStyle(h3);
-    const userStyle = getComputedStyle(userTurn);
-    const userOutlineStyle = getComputedStyle(userTurn, "::before");
-    const quoteStyle = getComputedStyle(quote);
-    return {
-      documentWidth: responseRect.width,
-      leftAxisDelta: responseRect.left - userTurnRect.left,
-      userLabels: userTurn.querySelectorAll(".turn-user-label").length,
-      userBackground: userStyle.backgroundColor,
-      userAccentRule: Number.parseFloat(userStyle.borderLeftWidth),
-      userRightBorder: Number.parseFloat(userStyle.borderRightWidth),
-      userOutline: userOutlineStyle.backgroundImage,
-      proseWidth: proseRect.width,
-      cardWidth: card.getBoundingClientRect().width,
-      wideTableWidth: wideTable.getBoundingClientRect().width,
-      h1Size: Number.parseFloat(h1Style.fontSize),
-      h2Size: Number.parseFloat(h2Style.fontSize),
-      h3Size: Number.parseFloat(h3Style.fontSize),
-      h1Divider: Number.parseFloat(h1Style.borderBottomWidth),
-      h2Marker: Number.parseFloat(h2Style.borderLeftWidth),
-      quoteRule: Number.parseFloat(quoteStyle.borderLeftWidth),
-      quoteBackground: quoteStyle.backgroundColor,
-      proseBackground: getComputedStyle(prose).backgroundColor,
-      codeOwnsOverflow: code.scrollWidth > code.clientWidth,
-      markdownTableTabIndex: markdownTable.tabIndex,
-      pageOverflow:
-        document.documentElement.scrollWidth - document.documentElement.clientWidth,
-    };
-  });
-}
-
-function readResponsiveDocumentLayout(p: Page) {
-  return p.evaluate(() => {
-    const zone = document.querySelector(".render-zone") as HTMLElement | null;
-    const response = document.querySelector(".response-document") as HTMLElement | null;
-    const prose = response?.querySelector(".turn-assistant") as HTMLElement | null;
-    const code = document.querySelector(
-      ".response-document .markdown pre code.hljs",
-    ) as HTMLElement | null;
-    const markdownTable = document.querySelector(
-      ".response-document .markdown-table-scroll",
-    ) as HTMLElement | null;
-    const richTable = document.querySelector(
-      ".response-document .rc-table",
-    ) as HTMLElement | null;
-    if (!zone || !response || !prose || !code || !markdownTable || !richTable) {
-      return null;
-    }
-    const zoneRect = zone.getBoundingClientRect();
-    const responseRect = response.getBoundingClientRect();
-    const proseRect = prose.getBoundingClientRect();
-    const markdownTableRect = markdownTable.getBoundingClientRect();
-    const richTableRect = richTable.getBoundingClientRect();
-    const zoneStyle = getComputedStyle(zone);
-    const contentLeft = zoneRect.left + Number.parseFloat(zoneStyle.paddingLeft);
-    const contentRight = zoneRect.right - Number.parseFloat(zoneStyle.paddingRight);
-    const fileContent = document.querySelector(".files-view .fv-content") as HTMLElement | null;
-    const filePanel = document.querySelector(".files-panel") as HTMLElement | null;
-    const fileContentRect = fileContent?.getBoundingClientRect();
-    const filePanelRect = filePanel?.getBoundingClientRect();
-    return {
-      documentWidth: responseRect.width,
-      proseWidth: proseRect.width,
-      proseHeight: proseRect.height,
-      pageOverflow:
-        document.documentElement.scrollWidth - document.documentElement.clientWidth,
-      documentContained:
-        responseRect.left >= contentLeft - 1 && responseRect.right <= contentRight + 1,
-      markdownTableContained:
-        markdownTableRect.left >= responseRect.left - 1 &&
-        markdownTableRect.right <= responseRect.right + 1,
-      richTableContained:
-        richTableRect.left >= responseRect.left - 1 &&
-        richTableRect.right <= responseRect.right + 1,
-      codeOwnsOverflow: code.scrollWidth > code.clientWidth,
-      fileContentContained:
-        !fileContentRect ||
-        !filePanelRect ||
-        (fileContentRect.left >= filePanelRect.left - 1 &&
-          fileContentRect.right <= filePanelRect.right + 1),
-      scrollTop: zone.scrollTop,
-      bottomDistance: zone.scrollHeight - zone.clientHeight - zone.scrollTop,
-    };
-  });
-}
 
 test("no token → 403, nothing served", async () => {
   const lockedOut = await browser.newPage();
@@ -323,194 +160,6 @@ test("?token= mints the cookie, cleans the URL, boots the shell", async () => {
   const cookies = await page.context().cookies(base);
   assert.ok(cookies.some((c) => c.name === "mirafold_token" && c.value === TOKEN));
   assert.equal(await page.locator(".fleet-title").textContent(), "Mirafold");
-});
-
-test("provider completions open before submit, transcript click focuses, and settled activity compacts", async () => {
-  const token = "e2e-native-prompt-9c2f";
-  await withFreshMockSession(
-    token,
-    async (page2) => {
-      const prompt = page2.locator(".prompt-box textarea");
-      const transcript = page2.locator(".render-zone");
-
-      // A trigger typed into page chrome is moved into the prompt and paints
-      // the provider catalog without sending a turn.
-      // Constrain the real listbox to force overflow with the mock catalog;
-      // this exercises the same keyboard scroll path as a long live catalog.
-      await page2.addStyleTag({
-        content: ".prompt-options { max-height: 48px !important; }",
-      });
-      await transcript.focus();
-      await page2.keyboard.press("/");
-      await page2.locator(".prompt-options").waitFor();
-      assert.equal(await prompt.inputValue(), "/");
-      // The Codex catalog: exactly the commands the adapter intercepts itself.
-      assert.equal(await page2.locator(".prompt-options [role=option]").count(), 2);
-      assert.equal(await page2.locator(".prompt-option-value", { hasText: "/model" }).count(), 1);
-      assert.equal(await page2.locator(".prompt-option-value", { hasText: "/effort" }).count(), 1);
-      assert.equal(await page2.locator(".turn-user").count(), 0, "opening a catalog submitted a turn");
-
-      await page2.keyboard.press("Escape");
-      await page2.locator(".prompt-options").waitFor({ state: "detached" });
-      await prompt.fill("");
-      await transcript.focus();
-      await page2.keyboard.press("$");
-      await page2.locator(".prompt-options").waitFor();
-      assert.equal(await page2.locator(".prompt-options [role=option]").count(), 2);
-      await page2.keyboard.press("ArrowDown");
-      const menuVisibility = await page2
-        .locator('.prompt-options [role="option"][aria-selected="true"]')
-        .evaluate((active) => {
-          const menu = active.parentElement!;
-          const activeRect = active.getBoundingClientRect();
-          const menuRect = menu.getBoundingClientRect();
-          return {
-            scrollTop: menu.scrollTop,
-            fullyVisible:
-              activeRect.top >= menuRect.top - 1 && activeRect.bottom <= menuRect.bottom + 1,
-          };
-        });
-      assert.ok(menuVisibility.scrollTop > 0, "keyboard selection did not scroll the listbox");
-      assert.equal(menuVisibility.fullyVisible, true, "active completion stayed offscreen");
-      await page2.keyboard.type("n");
-      assert.equal(await page2.locator(".prompt-options [role=option]").count(), 1);
-      assert.equal(await page2.locator(".prompt-option-value").textContent(), "$next");
-      assert.equal(
-        await page2.locator(".prompt-option-source").textContent(),
-        "Mirafold demo",
-        "catalog metadata must be visibly attributed inside trusted prompt chrome",
-      );
-      await page2.keyboard.press("Tab");
-      assert.equal(await prompt.inputValue(), "$next ");
-      assert.equal(await page2.locator(".turn-user").count(), 0, "Tab completion submitted a turn");
-
-      // Moving focus away and typing another trigger must preserve the draft,
-      // not replace it while routing the keystroke back to the composer.
-      await transcript.focus();
-      await page2.keyboard.press("$");
-      assert.equal(await prompt.inputValue(), "$next $");
-
-      // Shift+Escape was provisional and is deliberately gone. A plain
-      // desktop click on inert transcript chrome is the accepted return path.
-      await transcript.focus();
-      await page2.keyboard.press("Shift+Escape");
-      assert.equal(await transcript.evaluate((el) => document.activeElement === el), true);
-      await transcript.click({ position: { x: 2, y: 2 } });
-      assert.equal(
-        await page2.evaluate(() => document.activeElement?.matches(".prompt-box textarea")),
-        true,
-      );
-
-      // Successful provider activity becomes one terminal-sized record only
-      // when the turn settles. A failure remains visible at top level, and
-      // narration between commands (Codex's cadence) rides inside the fold
-      // instead of shattering it into singletons.
-      await prompt.fill("show transcript compact tool activity");
-      await prompt.press("Enter");
-      await page2.locator(".stop-btn").waitFor();
-      await page2.locator(".stop-btn").waitFor({ state: "detached" });
-      assert.equal(await page2.locator(".tool-activity-group").count(), 1);
-      // The fold's count speaks of ACTIONS only — absorbed narration
-      // (thinking rows riding inside the fold) must never inflate it.
-      assert.match(
-        await page2.locator(".tool-activity-label").innerText(),
-        /worked · 2 actions/,
-        "the fold label must count tool calls only, not absorbed narration",
-      );
-      assert.equal(await page2.locator(".tool-group").count(), 1);
-      assert.match(await page2.locator(".tool-group").textContent() ?? "", /No matching test file/);
-      assert.equal(
-        await page2.locator(".thinking-block", { hasText: "Weighing which check" }).count(),
-        0,
-        "interleaved narration leaked outside the settled fold",
-      );
-      await page2.locator(".tool-activity-head").click();
-      assert.equal(
-        await page2.evaluate(() => document.activeElement?.classList.contains("tool-activity-head")),
-        true,
-        "a transcript control click was redirected to the prompt",
-      );
-      assert.equal(await page2.locator(".tool-activity-calls .tool-block").count(), 2);
-      assert.equal(
-        await page2.locator(".tool-activity-calls .thinking-block", { hasText: "Weighing which check" }).count(),
-        1,
-        "the fold's expansion must replay the interleaved narration in place",
-      );
-
-      // Event delegation must treat ordinary transcript links as controls too.
-      // The real click path (pointerdown → pointerup → click) must leave focus
-      // on the link instead of redirecting it into the prompt.
-      const transcriptLink = page2.locator(".turn-assistant").last().locator("a").last();
-      await page2.locator(".turn-assistant").last().evaluate((el) => {
-        const link = document.createElement("a");
-        link.id = "transcript-control-fixture";
-        link.href = "#transcript-control-fixture";
-        link.textContent = "transcript link";
-        el.append(" ", link);
-      });
-      await prompt.evaluate((element) => {
-        const textarea = element as HTMLTextAreaElement;
-        const nativeFocus = textarea.focus.bind(textarea);
-        textarea.focus = (options?: FocusOptions) => {
-          textarea.dataset.transcriptLinkFocusCalls = String(
-            Number(textarea.dataset.transcriptLinkFocusCalls ?? "0") + 1,
-          );
-          nativeFocus(options);
-        };
-      });
-      await transcriptLink.click();
-      assert.equal(
-        await prompt.getAttribute("data-transcript-link-focus-calls"),
-        null,
-        "a transcript link click invoked prompt focus before restoring link focus",
-      );
-      assert.equal(
-        await transcriptLink.evaluate((el) => document.activeElement === el),
-        true,
-        "a transcript link click was redirected to the prompt",
-      );
-
-      // A text-selection gesture ends with the same pointerup event as a
-      // click. A live selection must win so copying transcript text remains
-      // possible and prompt focus does not collapse it.
-      const selected = await page2.locator(".turn-assistant").last().evaluate((el) => {
-        const text = document.createTreeWalker(el, NodeFilter.SHOW_TEXT).nextNode();
-        if (!text || !text.textContent) return { text: "", promptFocused: false };
-        const range = document.createRange();
-        range.setStart(text, 0);
-        range.setEnd(text, Math.min(12, text.textContent.length));
-        const selection = window.getSelection()!;
-        selection.removeAllRanges();
-        selection.addRange(range);
-        el.dispatchEvent(
-          new PointerEvent("pointerup", {
-            bubbles: true,
-            pointerType: "mouse",
-            button: 0,
-            isPrimary: true,
-          }),
-        );
-        return {
-          text: selection.toString(),
-          promptFocused: document.activeElement?.matches(".prompt-box textarea") ?? false,
-        };
-      });
-      assert.ok(selected.text.length > 0, "selection fixture did not select transcript text");
-      assert.equal(selected.promptFocused, false, "text selection was collapsed into prompt focus");
-      await page2.evaluate(() => window.getSelection()?.removeAllRanges());
-
-      // Touch must never summon the phone keyboard by focusing the prompt.
-      await transcript.focus();
-      await transcript.dispatchEvent("pointerup", {
-        pointerType: "touch",
-        button: 0,
-        isPrimary: true,
-      });
-      assert.equal(await transcript.evaluate((el) => document.activeElement === el), true);
-      await assertAxeClean(page2, "transcript click-to-focus");
-    },
-    "Codex",
-  );
 });
 
 test("the agent picker flexes to the window — no internal scrollbar through the squeeze ramp", async () => {
@@ -721,581 +370,6 @@ test("onboarding → a full mock turn renders in the DOM", async () => {
   await page.waitForSelector("text=Plan complete — all four steps done.", { timeout: 30_000 });
 });
 
-test("LD.1: a response document streams around a stable painting and soft shell interruption", async () => {
-  await withFreshMockSession("live-document-structure-7fd1", async (p) => {
-    await p.locator("textarea").click();
-    await p.keyboard.type("live document demo");
-    await p.keyboard.press("Enter");
-
-    await p.waitForSelector(".activity-line", { timeout: 15_000 });
-    const firstDocument = p.locator(".response-document").first();
-    await firstDocument
-      .locator(".turn-assistant", { hasText: "first section is already visible" })
-      .waitFor({ timeout: 15_000 });
-    assert.equal(
-      await p.locator(".activity-line").count(),
-      1,
-      "the first document content waited until turn_end",
-    );
-
-    const checkpoint = firstDocument.locator(".rc-card", {
-      hasText: "Live document checkpoint",
-    });
-    await checkpoint.waitFor({ timeout: 15_000 });
-    assert.equal(
-      await p.locator(".activity-line").count(),
-      1,
-      "the registry component waited until turn_end",
-    );
-    assert.equal(
-      await p.locator(".turn-assistant", { hasText: "Later prose begins beneath" }).count(),
-      0,
-      "the registry component did not paint before subsequent prose",
-    );
-    await checkpoint.evaluate((element) => {
-      (element as HTMLElement & { __liveDocumentIdentity?: string }).__liveDocumentIdentity =
-        "mounted-before-later-prose";
-    });
-
-    const notice = p.locator(".notice-line", {
-      hasText: "shell remains distinct",
-    });
-    await notice.waitFor({ timeout: 15_000 });
-    await p
-      .locator(".turn-assistant", { hasText: "Later prose begins beneath" })
-      .waitFor({ timeout: 15_000 });
-
-    const documents = p.locator(".response-document");
-    await p.waitForFunction(
-      () => document.querySelectorAll(".response-document").length >= 2,
-    );
-    const documentMeta = await documents.evaluateAll((elements) =>
-      elements.slice(0, 2).map((element) => ({
-        responseKey: element.getAttribute("data-response-key"),
-        continuation: element.hasAttribute("data-response-continuation"),
-      })),
-    );
-    assert.equal(documentMeta[0]?.responseKey, documentMeta[1]?.responseKey);
-    assert.equal(documentMeta[0]?.continuation, false);
-    assert.equal(documentMeta[1]?.continuation, true);
-
-    assert.equal(
-      await p.evaluate(() => {
-        const first = document.querySelectorAll(".response-document")[0];
-        const shell = document.querySelector(".notice-line");
-        const second = document.querySelectorAll(".response-document")[1];
-        if (!first || !shell || !second) return false;
-        return Boolean(
-          first.compareDocumentPosition(shell) & Node.DOCUMENT_POSITION_FOLLOWING,
-        ) && Boolean(
-          shell.compareDocumentPosition(second) & Node.DOCUMENT_POSITION_FOLLOWING,
-        );
-      }),
-      true,
-      "the soft shell row moved out of transcript order",
-    );
-    assert.equal(
-      await checkpoint.evaluate(
-        (element) =>
-          (element as HTMLElement & { __liveDocumentIdentity?: string })
-            .__liveDocumentIdentity,
-      ),
-      "mounted-before-later-prose",
-      "later prose remounted the earlier registry component",
-    );
-
-    const secondDocument = documents.nth(1);
-    await secondDocument
-      .locator(".rc-table", { hasText: "Composition checks" })
-      .waitFor({ timeout: 15_000 });
-    await secondDocument
-      .locator(".turn-assistant", { hasText: "response finished as one live composition" })
-      .waitFor({ timeout: 15_000 });
-    const secondKinds = await secondDocument.locator(":scope > *").evaluateAll((elements) =>
-      elements.map((element) =>
-        element.classList.contains("turn-assistant")
-          ? "text"
-          : element.classList.contains("turn-render")
-            ? "render"
-            : "other",
-      ),
-    );
-    assert.deepEqual(secondKinds, ["text", "render", "text"]);
-    await p.waitForSelector(".activity-line", { state: "detached", timeout: 15_000 });
-  });
-});
-
-test("LD.2: document measure, hierarchy, rich lane, local overflow, and short-answer restraint", async () => {
-  await withFreshMockSession("live-document-visual-2a91", async (p) => {
-    await p.setViewportSize({ width: 1440, height: 1000 });
-    const prompt = await settleLiveDocument(p);
-
-    const metrics = await readLiveDocumentPresentation(p);
-
-    assert.ok(metrics, "the canonical document fixture did not fully render");
-    assert.ok(metrics.documentWidth > 900 && metrics.documentWidth <= 1160.5);
-    assert.ok(
-      Math.abs(metrics.leftAxisDelta) <= 1,
-      `the response moved off the transcript's left axis by ${metrics.leftAxisDelta}px`,
-    );
-    assert.equal(metrics.userLabels, 0);
-    assert.notEqual(metrics.userBackground, "rgba(0, 0, 0, 0)");
-    assert.ok(metrics.userAccentRule >= 3);
-    assert.equal(metrics.userRightBorder, 0);
-    assert.match(metrics.userOutline, /linear-gradient/);
-    assert.ok(metrics.proseWidth < metrics.documentWidth);
-    assert.ok(metrics.cardWidth < metrics.documentWidth);
-    assert.ok(Math.abs(metrics.wideTableWidth - metrics.documentWidth) <= 1);
-    assert.ok(metrics.h1Size > metrics.h2Size && metrics.h2Size > metrics.h3Size);
-    assert.ok(metrics.h1Divider >= 1);
-    assert.ok(metrics.h2Marker >= 2);
-    assert.ok(metrics.quoteRule >= 2);
-    assert.notEqual(metrics.quoteBackground, "rgba(0, 0, 0, 0)");
-    assert.equal(metrics.proseBackground, "rgba(0, 0, 0, 0)");
-    assert.equal(metrics.codeOwnsOverflow, true);
-    assert.equal(metrics.markdownTableTabIndex, 0);
-    assert.ok(metrics.pageOverflow <= 1);
-    await assertAxeClean(p, "live document composition");
-
-    await prompt.fill("one sentence document");
-    await prompt.press("Enter");
-    const shortDocument = p.locator(".response-document").last();
-    await shortDocument
-      .locator(".turn-assistant", { hasText: "workspace is ready for the next decision" })
-      .waitFor({ timeout: 15_000 });
-    await p.locator(".activity-line").waitFor({ state: "detached", timeout: 15_000 });
-    const shortShape = await shortDocument.evaluate((element) => {
-      const prose = element.querySelector(".turn-assistant") as HTMLElement | null;
-      const style = prose ? getComputedStyle(prose) : null;
-      return {
-        children: element.children.length,
-        paintings: element.querySelectorAll(".rc, .artifact").length,
-        background: style?.backgroundColor,
-        borderWidth: style?.borderWidth,
-      };
-    });
-    assert.deepEqual(shortShape, {
-      children: 1,
-      paintings: 0,
-      background: "rgba(0, 0, 0, 0)",
-      borderWidth: "0px",
-    });
-  });
-});
-
-test("LD.3: response documents reflow across Explorer, file view, pin dock, and narrow desktop", async () => {
-  await withFreshMockSession("live-document-responsive-18c4", async (p) => {
-    await p.setViewportSize({ width: 1440, height: 1000 });
-    const prompt = await settleLiveDocument(p);
-
-    const measure = () => readResponsiveDocumentLayout(p);
-
-    const center = await measure();
-    assert.ok(center, "center-only response metrics are missing");
-
-    await p.locator(".ab-files").click();
-    await p.waitForSelector(".files-panel .files-row");
-    const explorer = await measure();
-    assert.ok(explorer, "Explorer response metrics are missing");
-
-    const checkpointTurn = p.locator(".turn-render", {
-      has: p.locator(".rc-card", { hasText: "Live document checkpoint" }),
-    });
-    await checkpointTurn.hover();
-    await checkpointTurn.locator(".pin-btn").click();
-    await p.waitForSelector(".pin-dock .rc-card");
-    const both = await measure();
-    assert.ok(both, "Explorer plus pin-dock response metrics are missing");
-
-    await p.locator(".files-file-row", { hasText: "package.json" }).first().click();
-    await p.waitForSelector(".files-view .fv-content");
-    const fileAndDock = await measure();
-    assert.ok(fileAndDock, "file-view plus pin-dock response metrics are missing");
-
-    await p.setViewportSize({ width: 980, height: 760 });
-    await p.waitForFunction(() => window.innerWidth === 980);
-    const narrow = await measure();
-    assert.ok(narrow, "narrow three-pane response metrics are missing");
-
-    for (const [name, metrics] of [
-      ["center", center],
-      ["Explorer", explorer],
-      ["Explorer + dock", both],
-      ["file view + dock", fileAndDock],
-      ["narrow three-pane", narrow],
-    ] as const) {
-      assert.ok(metrics.pageOverflow <= 1, `${name} overflowed the page by ${metrics.pageOverflow}px`);
-      assert.equal(metrics.documentContained, true, `${name} document escaped the center column`);
-      assert.equal(metrics.markdownTableContained, true, `${name} Markdown table escaped its document`);
-      assert.equal(metrics.richTableContained, true, `${name} registry table escaped its document`);
-      assert.equal(metrics.fileContentContained, true, `${name} file content escaped Explorer`);
-    }
-    assert.equal(center.codeOwnsOverflow, true, "wide code did not retain local overflow");
-    assert.ok(center.documentWidth > explorer.documentWidth);
-    assert.ok(explorer.documentWidth > both.documentWidth);
-    assert.ok(Math.abs(fileAndDock.documentWidth - both.documentWidth) <= 1);
-    assert.ok(both.documentWidth > narrow.documentWidth);
-    assert.ok(narrow.documentWidth >= 280, `narrow document collapsed to ${narrow.documentWidth}px`);
-    assert.ok(narrow.proseWidth < center.proseWidth);
-    assert.ok(narrow.proseHeight > center.proseHeight, "narrow prose did not reflow vertically");
-
-    // Exercise the remaining intrinsically wide content while all three
-    // desktop columns are present at 980px. The artifact deliberately has a
-    // 720px internal canvas: its own iframe must scroll, never the workbench.
-    await p.locator(".render-zone").evaluate((element) => {
-      element.scrollTop = element.scrollHeight;
-    });
-    await submitPrompt(p, "responsive document stress");
-    const stressDocument = p.locator(".response-document", { hasText: "Width stress" });
-    await stressDocument.locator(".rc-diff").waitFor({ timeout: 15_000 });
-    await stressDocument.locator(".rc-chart").waitFor({ timeout: 15_000 });
-    const finalStressProse = stressDocument.locator(".turn-assistant", {
-      hasText: "without widening the workbench",
-    });
-    await stressDocument.locator(".artifact").waitFor({ timeout: 15_000 });
-    assert.equal(
-      await p.locator(".activity-line").count(),
-      1,
-      "the artifact waited until turn_end",
-    );
-    assert.equal(
-      await finalStressProse.count(),
-      0,
-      "the artifact did not paint before subsequent prose",
-    );
-    const artifactHandle = await stressDocument
-      .locator("iframe.artifact-frame")
-      .elementHandle({ timeout: 15_000 });
-    assert.ok(artifactHandle, "responsive artifact frame did not render");
-    const artifactFrame = await artifactHandle.contentFrame();
-    assert.ok(artifactFrame, "responsive artifact frame was unavailable");
-    await artifactFrame.waitForSelector(".wide", { timeout: 15_000 });
-    await finalStressProse.waitFor({ timeout: 15_000 });
-    await p.locator(".activity-line").waitFor({ state: "detached", timeout: 15_000 });
-
-    const stressShape = await stressDocument.evaluate((element) => {
-      const documentRect = element.getBoundingClientRect();
-      const chartPlot = element.querySelector(".rc-chart-plot") as HTMLElement | null;
-      const chartCanvas = chartPlot?.querySelector("svg") as SVGElement | null;
-      const paintings = [
-        element.querySelector(".rc-diff"),
-        element.querySelector(".rc-chart"),
-        element.querySelector(".artifact"),
-        element.querySelector("iframe.artifact-frame"),
-      ];
-      return {
-        allContained: paintings.every((painting) => {
-          if (!(painting instanceof HTMLElement)) return false;
-          const rect = painting.getBoundingClientRect();
-          return rect.left >= documentRect.left - 1 && rect.right <= documentRect.right + 1;
-        }),
-        pageOverflow:
-          document.documentElement.scrollWidth - document.documentElement.clientWidth,
-        chartOwnsOverflow:
-          chartPlot ? chartPlot.scrollWidth > chartPlot.clientWidth : false,
-        chartTabIndex: chartPlot?.tabIndex ?? -1,
-        chartCanvasWidth: chartCanvas?.getBoundingClientRect().width ?? 0,
-      };
-    });
-    assert.equal(stressShape.allContained, true, "wide painting escaped the narrow document");
-    assert.ok(stressShape.pageOverflow <= 1);
-    assert.equal(stressShape.chartOwnsOverflow, true, "narrow chart did not own its overflow");
-    assert.equal(stressShape.chartTabIndex, 0, "narrow chart overflow is not keyboard reachable");
-    assert.ok(stressShape.chartCanvasWidth >= 599, "chart canvas shrank below readable measure");
-    assert.equal(
-      await artifactFrame.evaluate(
-        () => document.documentElement.scrollWidth > document.documentElement.clientWidth,
-      ),
-      true,
-      "wide artifact content did not keep its overflow inside the sandbox",
-    );
-    const tail = await measure();
-    assert.ok(tail, "tail metrics are missing after responsive stress");
-    assert.ok(tail.bottomDistance <= 2, `streaming stopped ${tail.bottomDistance}px above the tail`);
-
-    // A user who has scrolled back must not be snapped to the tail merely
-    // because workspace chrome closes and the document reflows wider.
-    const manualScroll = await p.locator(".render-zone").evaluate((element) => {
-      element.scrollTop = 120;
-      return element.scrollTop;
-    });
-    assert.ok(manualScroll > 0, "fixture is not tall enough to exercise manual scroll");
-    await p.locator(".ab-files").click();
-    await p.waitForSelector(".files-panel", { state: "detached" });
-    const dockOnly = await measure();
-    assert.ok(dockOnly, "dock-only response metrics are missing");
-    assert.ok(dockOnly.documentWidth > narrow.documentWidth);
-    assert.ok(
-      Math.abs(dockOnly.scrollTop - manualScroll) <= 2,
-      `closing Explorer moved manual scroll from ${manualScroll}px to ${dockOnly.scrollTop}px`,
-    );
-
-    await p.locator(".pin-dock .dock-btn").click();
-    await p.waitForSelector(".pin-tab");
-    const collapsedDock = await measure();
-    assert.ok(collapsedDock, "collapsed-dock response metrics are missing");
-    assert.ok(collapsedDock.documentWidth > dockOnly.documentWidth);
-    assert.ok(collapsedDock.pageOverflow <= 1);
-
-    await p.locator(".pin-stub", { hasText: "card" }).click();
-    await p.waitForSelector(".pin-tab", { state: "detached" });
-    const restored = await measure();
-    assert.ok(restored, "restored response metrics are missing");
-    assert.ok(restored.documentWidth > collapsedDock.documentWidth);
-    assert.ok(restored.pageOverflow <= 1);
-  });
-});
-
-test("LD.4: long text, focus, announcements, reduced motion, and every theme remain restrained", async () => {
-  await withFreshMockSession("live-document-closure-63bc", async (p) => {
-    await p.setViewportSize({ width: 1440, height: 800 });
-    await p.emulateMedia({ reducedMotion: "reduce" });
-    const prompt = await submitPrompt(p, "document closure stress");
-
-    const response = p.locator(".response-document", {
-      hasText: "This heading-free technical note",
-    });
-    await response.locator("text=Closure stress complete.").waitFor({ timeout: 30_000 });
-    await p.locator(".activity-line").waitFor({ state: "detached", timeout: 15_000 });
-
-    const shape = await response.evaluate((element) => {
-      const zone = document.querySelector(".render-zone") as HTMLElement | null;
-      const prose = element.querySelector(".turn-assistant") as HTMLElement | null;
-      const code = element.querySelector("pre code.hljs") as HTMLElement | null;
-      const style = prose ? getComputedStyle(prose) : null;
-      return {
-        children: element.children.length,
-        headings: element.querySelectorAll("h1, h2, h3, h4, h5, h6").length,
-        paintings: element.querySelectorAll(".rc, .artifact").length,
-        height: element.getBoundingClientRect().height,
-        viewportHeight: window.innerHeight,
-        proseOwnsWidth: prose ? prose.scrollWidth <= prose.clientWidth + 1 : false,
-        codeOwnsOverflow: code ? code.scrollWidth > code.clientWidth : false,
-        codeTabIndex: code?.tabIndex ?? -1,
-        pageOverflow:
-          document.documentElement.scrollWidth - document.documentElement.clientWidth,
-        bottomDistance: zone ? zone.scrollHeight - zone.clientHeight - zone.scrollTop : -1,
-        responseRole: element.getAttribute("role"),
-        responseLive: element.getAttribute("aria-live"),
-        proseAnimation: style?.animationName,
-        proseTransition: style?.transitionDuration,
-      };
-    });
-    assert.equal(shape.children, 1);
-    assert.equal(shape.headings, 0);
-    assert.equal(shape.paintings, 0);
-    assert.ok(shape.height > shape.viewportHeight * 2, "the long-answer fixture is not genuinely tall");
-    assert.equal(shape.proseOwnsWidth, true);
-    assert.equal(shape.codeOwnsOverflow, true);
-    assert.equal(shape.codeTabIndex, 0);
-    assert.ok(shape.pageOverflow <= 1);
-    assert.ok(shape.bottomDistance <= 2);
-    assert.equal(shape.responseRole, null);
-    assert.equal(shape.responseLive, null);
-    assert.equal(shape.proseAnimation, "none");
-    assert.equal(shape.proseTransition, "0s");
-    assert.equal(await response.locator('a[href^="https://example.invalid/diagnostics/"]').count(), 1);
-    assert.equal(
-      await response.locator("code", { hasText: "deeply-nested-workspace" }).count(),
-      1,
-    );
-
-    const log = p.locator('.render-zone[role="log"]');
-    assert.equal(await log.getAttribute("aria-live"), "off");
-    await p.waitForFunction(
-      () =>
-        document
-          .querySelector('[role="status"]')
-          ?.textContent?.includes("Response truncated for reading"),
-      undefined,
-      { timeout: 15_000 },
-    );
-
-    await p.evaluate(() => (document.activeElement as HTMLElement | null)?.blur());
-    const selection = await response.locator(".turn-assistant").evaluate((element) => {
-      const text = document.createTreeWalker(element, NodeFilter.SHOW_TEXT).nextNode();
-      if (!text || !text.textContent) return { text: "", promptFocused: false };
-      const range = document.createRange();
-      range.setStart(text, 0);
-      range.setEnd(text, Math.min(24, text.textContent.length));
-      const current = window.getSelection()!;
-      current.removeAllRanges();
-      current.addRange(range);
-      element.dispatchEvent(
-        new PointerEvent("pointerup", {
-          bubbles: true,
-          pointerType: "mouse",
-          button: 0,
-          isPrimary: true,
-        }),
-      );
-      return {
-        text: current.toString(),
-        promptFocused: document.activeElement?.matches(".prompt-box textarea") ?? false,
-      };
-    });
-    assert.ok(selection.text.length > 0);
-    assert.equal(selection.promptFocused, false, "selection was collapsed into prompt focus");
-
-    await response.evaluate((element) => {
-      window.getSelection()?.removeAllRanges();
-      element.dispatchEvent(
-        new PointerEvent("pointerup", {
-          bubbles: true,
-          pointerType: "mouse",
-          button: 0,
-          isPrimary: true,
-        }),
-      );
-    });
-    assert.equal(
-      await prompt.evaluate((element) => document.activeElement === element),
-      true,
-      "a plain transcript click no longer focuses the prompt",
-    );
-
-    await response.evaluate((element) => {
-      (element as HTMLElement & { __ld4Identity?: string }).__ld4Identity = "before-themes";
-    });
-    const baseGeometry = await response.evaluate((element) => ({
-      left: element.getBoundingClientRect().left,
-      width: element.getBoundingClientRect().width,
-    }));
-    await p.locator(".sb-settings").click();
-    await p.locator(".settings-card").waitFor();
-    for (const theme of THEMES) {
-      const groupLabel = theme.appearance === "light" ? "Light themes" : "Dark themes";
-      const group = p.locator(`.theme-group[aria-label="${groupLabel}"]`);
-      const names = await group.locator(".theme-row-name").allInnerTexts();
-      const index = names.indexOf(theme.displayName);
-      assert.ok(index >= 0, `theme row ${theme.id} is missing`);
-      await group.locator(".theme-row").nth(index).click();
-      await p.waitForFunction(
-        (id) => document.documentElement.getAttribute("data-theme") === id,
-        theme.id,
-      );
-      const themed = await response.evaluate((element) => ({
-        identity: (element as HTMLElement & { __ld4Identity?: string }).__ld4Identity,
-        left: element.getBoundingClientRect().left,
-        width: element.getBoundingClientRect().width,
-        pageOverflow:
-          document.documentElement.scrollWidth - document.documentElement.clientWidth,
-        bodyBackground: getComputedStyle(document.body).backgroundColor,
-        proseColor: getComputedStyle(element.querySelector(".turn-assistant")!).color,
-      }));
-      assert.equal(themed.identity, "before-themes", `${theme.id} remounted the document`);
-      assert.ok(Math.abs(themed.left - baseGeometry.left) <= 1, `${theme.id} moved the document`);
-      assert.ok(Math.abs(themed.width - baseGeometry.width) <= 1, `${theme.id} changed document width`);
-      assert.ok(themed.pageOverflow <= 1, `${theme.id} introduced page overflow`);
-      assert.notEqual(themed.bodyBackground, "rgba(0, 0, 0, 0)");
-      assert.notEqual(themed.proseColor, "rgba(0, 0, 0, 0)");
-    }
-    await p.keyboard.press("Escape");
-    await p.locator(".settings-card").waitFor({ state: "detached" });
-    await assertAxeClean(p, "LD.4 closure stress");
-  });
-});
-
-test("SA.1: a parallel fan-out renders three live subagent decks, out of order, expandable", async () => {
-  await withFreshMockSession("sa1-cards-7fd1", async (p) => {
-    await p.locator("textarea").click();
-    await p.keyboard.type("delegate the research");
-    await p.keyboard.press("Enter");
-
-    // Three cards appear as their spawns land (350/430/510ms in the mock).
-    await p.waitForFunction(
-      () => document.querySelectorAll(".subagent-deck").length === 3,
-      undefined,
-      { timeout: 15_000 },
-    );
-
-    // While running: a live card ticks — elapsed seconds in the meta line and
-    // a current action in the live slot. Grab the slow "trace the token path"
-    // card, which runs the longest.
-    const tokenCard = p.locator(".subagent-deck", { hasText: "trace the token path" });
-    await p.waitForFunction(
-      () => {
-        const cards = [...document.querySelectorAll(".subagent-deck-running")];
-        return cards.some((c) => /·\s*\d+s/.test(c.querySelector(".subagent-deck-meta")?.textContent ?? ""));
-      },
-      undefined,
-      { timeout: 10_000 },
-    );
-
-    // OUT OF ORDER: "map session handling" (spawned second, fastest pace)
-    // finishes while "trace the token path" is still running.
-    await p.waitForFunction(
-      () => {
-        // No const-assigned arrows in here: esbuild's keepNames would inject
-        // a __name helper the browser page doesn't have.
-        const cards = [...document.querySelectorAll(".subagent-deck")];
-        const settled = cards.find((c) => c.textContent?.includes("map session handling"));
-        const slow = cards.find((c) => c.textContent?.includes("trace the token path"));
-        return (
-          settled?.classList.contains("subagent-deck-done") === true &&
-          slow?.classList.contains("subagent-deck-running") === true
-        );
-      },
-      undefined,
-      { timeout: 15_000 },
-    );
-
-    // Child tool churn must not steer the ROOT activity line — each deck
-    // shows its own current action (bughunt 2026-08-14 r2). The label may
-    // legitimately read the spawn state or the generic fallback (a FINISHED
-    // spawn clears its own name), but never a child's tool. Sampled across
-    // the still-running slow agent's ~640ms tool cadence so a pre-fix
-    // child-name label cannot slip between reads.
-    for (let sample = 0; sample < 8; sample++) {
-      const activityText = await p.locator(".activity-label").innerText();
-      assert.match(activityText, /^(Task|working)…$/);
-      await new Promise((resolve) => setTimeout(resolve, 150));
-    }
-
-    // The turn concludes; all three cards settle, elapsed stops being shown
-    // (client-side timing is only honest while live), result lines ride.
-    await p.waitForSelector("text=All three subagents reported back", { timeout: 30_000 });
-    assert.equal(await p.locator(".subagent-deck-done").count(), 3);
-    assert.equal(await p.locator(".subagent-deck-running").count(), 0);
-    const doneMeta = await tokenCard.locator(".subagent-deck-meta").innerText();
-    assert.doesNotMatch(doneMeta, /·\s*\d+s/);
-    assert.match(doneMeta, /4 tools/);
-    assert.match(doneMeta, /never leaves the daemon/);
-
-    // Expand → the nested calls are all there, AND the subagent's own words
-    // (SA.2: narration + reasoning, interleaved, inert plain text — the
-    // narration precedes the first tool row, true stream order).
-    await tokenCard.locator(".subagent-deck-head").click();
-    assert.equal(await tokenCard.locator(".subagent-calls .tool-block").count(), 4);
-    const prose = tokenCard.locator(".subagent-prose");
-    assert.match(await prose.first().innerText(), /Following the cookie from auth\.ts/);
-    assert.match(
-      await tokenCard.locator(".subagent-prose-thinking").innerText(),
-      /confirming the browser side/,
-    );
-    const expandedTexts = await tokenCard
-      .locator(".subagent-calls > *")
-      .evaluateAll((nodes) => nodes.map((n) => n.className));
-    assert.ok(
-      expandedTexts[0].includes("subagent-prose"),
-      "narration precedes the first tool row in stream order",
-    );
-    // The prose is NOT rendered as markdown — no <p>/<em> children, raw text.
-    assert.equal(await prose.first().locator("p, em, strong, a, code").count(), 0);
-    await tokenCard.locator(".subagent-deck-head").click();
-    assert.equal(await tokenCard.locator(".subagent-calls").count(), 0);
-
-    await assertAxeClean(p, "subagent decks");
-    await noSideScroll(p);
-
-    // Phone width: the same cards stack — no pane, no side scroll, drill-in
-    // untouched (the transcript is the same DOM at every width).
-    await p.setViewportSize({ width: 390, height: 844 });
-    assert.equal(await p.locator(".subagent-deck").count(), 3);
-    await noSideScroll(p);
-  });
-});
-
 test("A.1: announcer regions exist, spoke the turn, and the transcript is silent", async () => {
   // The two shell-owned announcer regions (Announcer.tsx): polite for turn
   // progress, assertive reserved for errors/permissions. `.sr-only` hides
@@ -1499,42 +573,6 @@ test("status-list component: one verdict pill per row, glyph + word, all five st
     assert.ok((await pill.innerText()).includes(status));
   }
   assert.match(await block.locator(".rc-status-row", { hasText: "relay probe" }).innerText(), /4007/);
-});
-
-test("diagram component: mermaid renders as SVG inside the sandbox; broken source shows itself", async () => {
-  // This test used to borrow the suite's long-lived session. Its failure was
-  // misattributed to Mermaid's lazy chunk, but the outer .rc-diagram never
-  // arrived: an earlier test could finish its DOM assertions just before its
-  // turn_end and leave this prompt contending with the one-follow-up gate.
-  // Own a session so this test reaches the renderer or fails for a renderer
-  // reason; the actual lazy chunk, iframe, CSP, postMessage, and parse error
-  // paths below remain production-real.
-  await withFreshMockSession("e2e-diagram-9c2f", async (page2) => {
-    await page2.waitForSelector("textarea");
-    await page2.locator("textarea").fill("diagram demo");
-    await page2.keyboard.press("Enter");
-
-    const block = page2.locator(".rc-diagram", { hasText: "Relay pairing flow" });
-    await block.waitFor({ timeout: 20_000 });
-    // The runtime is a lazy ~3.6 MB chunk — give the first render room.
-    const frame = block.locator("iframe.rc-diagram-frame");
-    await frame.waitFor({ timeout: 20_000 });
-    const svg = page2
-      .frameLocator(".rc-diagram:has-text('Relay pairing flow') iframe")
-      .locator("#host svg");
-    await svg.waitFor({ timeout: 20_000 });
-    // The frame reported its measured height back — the host sized to fit.
-    const h = await frame.evaluate((el) => el.getBoundingClientRect().height);
-    assert.ok(h > 130, `frame did not grow to the diagram (h=${h})`);
-    // The shell-drawn chrome badges the sandbox, outside the frame.
-    assert.match(await block.locator(".rc-diagram-badge").innerText(), /sandboxed/);
-
-    // Broken source: the failure state carries the message AND the source text.
-    const failed = page2.locator(".rc-diagram-failed", { hasText: "broken diagram" });
-    await failed.waitFor({ timeout: 20_000 });
-    assert.match(await failed.innerText(), /diagram didn't render/);
-    assert.ok((await failed.locator(".rc-diagram-source").innerText()).includes("nope"));
-  });
 });
 
 test("image component: a resolved shot renders as a real <img>; a refused one says why", async () => {
@@ -1961,160 +999,6 @@ test("settings card: gear opens it, picking applies live and writes the slot (S.
   await page.locator(".settings-backdrop").click({ position: { x: 8, y: 8 } });
   assert.equal(await page.locator(".settings-card").count(), 0);
   assert.equal(await dataTheme(), "dark"); // ends where the suite expects
-});
-
-test("NF: hidden viewport toasts a permission then the turn end; visibility closes both", async () => {
-  const token = "e2e-notify-7b31";
-  type ToastRec = { title: string; body?: string; tag?: string; closed: boolean };
-  const toasts = (p: Page) =>
-    p.evaluate(() => (window as unknown as { __TOASTS__: ToastRec[] }).__TOASTS__);
-  // A JS string, not a function — the accessor property would otherwise
-  // compile with the module-scope __name wrapper (same keepNames trap as the
-  // stub below) and die serialized.
-  const setVisibility = (p: Page, state: "hidden" | "visible") =>
-    p.evaluate(
-      `Object.defineProperty(document, "visibilityState", {
-         configurable: true, get: function () { return "${state}"; },
-       });
-       document.dispatchEvent(new Event("visibilitychange"));`,
-    );
-  await withFreshMockSession(
-    token,
-    async (page2) => {
-      // The toggle lives in the settings card and defaults off; flipping it
-      // writes the preference (the stub's grant is already "granted", so no
-      // permission dance here — that branch is Tier-1 logic).
-      await page2.locator(".sb-settings").click();
-      await page2.waitForSelector(".settings-card");
-      const row = page2.locator(".notify-row");
-      assert.equal(await row.getAttribute("aria-checked"), "false");
-      await row.click();
-      assert.equal(await row.getAttribute("aria-checked"), "true");
-      assert.equal(
-        await page2.evaluate(() => localStorage.getItem("mirafold-notify")),
-        "1",
-      );
-      await assertAxeClean(page2, "settings notifications section");
-      await page2.keyboard.press("Escape");
-      await page2.waitForSelector(".settings-card", { state: "detached" });
-
-      // Background the tab — the notifier reads visibilityState live, and
-      // only a hidden tab may toast.
-      await setVisibility(page2, "hidden");
-
-      // A permission lands while hidden → exactly one toast, session-tagged,
-      // shell-composed title with the engine's ask as inert text.
-      await page2.locator("textarea").click();
-      await page2.keyboard.type("run something dangerous");
-      await page2.keyboard.press("Enter");
-      await page2.waitForSelector(".perm-bar");
-      await page2.waitForFunction(
-        () => (window as unknown as { __TOASTS__?: unknown[] }).__TOASTS__?.length === 1,
-        undefined,
-        { timeout: 15_000 },
-      );
-      const [first] = await toasts(page2);
-      assert.match(first.title, /^⚠ permission — /);
-      assert.match(first.body ?? "", /Claude Code wants Bash: rm -rf/);
-      assert.ok(first.tag?.startsWith("mirafold-"), `tag was ${first.tag}`);
-
-      // Answering retires the permission toast (the state moved on); the
-      // finishing turn then toasts once more under the same tag.
-      await page2.locator(".perm-allow").click();
-      await page2.waitForFunction(
-        () => {
-          const t = (window as unknown as { __TOASTS__: { closed: boolean }[] }).__TOASTS__;
-          return t.length === 2 && t[0].closed;
-        },
-        undefined,
-        { timeout: 15_000 },
-      );
-      const [, second] = await toasts(page2);
-      assert.match(second.title, /^✓ turn finished — /);
-      assert.equal(second.tag, first.tag);
-
-      // Coming back closes everything this tab created — the page itself is
-      // now the notification.
-      await setVisibility(page2, "visible");
-      await page2.waitForFunction(() =>
-        (window as unknown as { __TOASTS__: { closed: boolean }[] }).__TOASTS__.every(
-          (t) => t.closed,
-        ),
-      );
-    },
-    "Claude Code",
-    async (p) => {
-      // Notification stubbed before boot: headless Chrome auto-denies the
-      // real API and an OS toast is invisible to the DOM anyway — the
-      // recorder IS the observable surface. A plain-JS STRING, not a
-      // function: tsx compiles this file with esbuild keepNames, which
-      // injects a module-scope `__name` helper into compiled classes —
-      // Playwright then serializes the function without the helper and the
-      // init script dies on a ReferenceError before installing the stub
-      // (diagnosed 2026-08-12 via pageerror probe).
-      await p.addInitScript(`
-        const spawned = [];
-        window.__TOASTS__ = spawned;
-        window.Notification = class {
-          constructor(title, opts) {
-            this.rec = { title, body: opts && opts.body, tag: opts && opts.tag, closed: false };
-            this.onclick = null;
-            spawned.push(this.rec);
-          }
-          close() { this.rec.closed = true; }
-          static get permission() { return "granted"; }
-          static requestPermission() { return Promise.resolve("granted"); }
-        };
-      `);
-    },
-  );
-});
-
-test("FD: a dropped file uploads, stages on disk, and its quoted path lands in the prompt", async () => {
-  const token = "e2e-filedrop-4a19";
-  await withFreshMockSession(token, async (page2) => {
-    // The drag listeners attach only once the session is attached
-    // (meta.sessionId) — a dispatch racing the mount fires into the void,
-    // so wait for the session UI first (diagnosed 2026-08-12 by probe).
-    await page2.waitForSelector(".prompt-box textarea");
-    // Synthesize a real file drag: DataTransfer + File are native in
-    // Chromium. A JS string, not a function — the keepNames __name trap
-    // (see the NF test above) breaks serialized closures.
-    await page2.evaluate(`
-      window.__DT__ = (() => {
-        const dt = new DataTransfer();
-        dt.items.add(new File(["drag and drop payload snowman"], "dropped notes.txt", { type: "text/plain" }));
-        return dt;
-      })();
-      document.body.dispatchEvent(new DragEvent("dragenter", { bubbles: true, cancelable: true, dataTransfer: window.__DT__ }));
-    `);
-    // Mid-drag, the shell-owned overlay invites the drop.
-    await page2.waitForSelector(".drop-overlay");
-    await page2.evaluate(`
-      document.body.dispatchEvent(new DragEvent("drop", { bubbles: true, cancelable: true, dataTransfer: window.__DT__ }));
-    `);
-    // The staged path lands in the prompt, quoted (the name carries a space).
-    await page2.waitForFunction(
-      () => {
-        const el = document.querySelector(".prompt-box textarea") as HTMLTextAreaElement | null;
-        return el?.value.includes("mirafold-uploads") ?? false;
-      },
-      undefined,
-      { timeout: 15_000 },
-    );
-    assert.equal(await page2.locator(".drop-overlay").count(), 0, "overlay clears on drop");
-    const value = await page2.locator(".prompt-box textarea").inputValue();
-    const m = value.match(/'([^']*mirafold-uploads[^']*)'/);
-    assert.ok(m, `expected a quoted staged path in the prompt, got: ${value}`);
-    const staged = m![1];
-    assert.match(staged, /dropped notes\.txt$/);
-    // The path is honest: the exact dropped bytes sit at it on disk.
-    assert.equal(readFileSync(staged, "utf8"), "drag and drop payload snowman");
-    // The attach announced politely (the screen-reader path).
-    const status = await page2.locator('[role="status"]').innerText();
-    assert.match(status, /Attached dropped notes\.txt/);
-    rmSync(path.dirname(staged), { recursive: true, force: true });
-  });
 });
 
 test("Esc on an open card is exclusive — it dismisses without interrupting the turn", async () => {
@@ -2562,318 +1446,7 @@ test("status bar: new sits beside home, end is the far-right control, ?new opens
   await fresh.close();
 });
 
-test("streaming holds a scrolled-up reader in place, and re-follows once back at the bottom", async () => {
-  // Own the transcript and make it scrollable at a fixed viewport. The old
-  // version borrowed 40 prior tests' worth of content, then jumped to the
-  // bottom programmatically and sampled once; neither is a user's re-arm
-  // path, and the delayed scroll event raced the next streamed paint.
-  await withFreshMockSession("e2e-follow-tail-9c2f", async (page2) => {
-    await page2.setViewportSize({ width: 900, height: 520 });
-    await page2.waitForSelector("textarea");
-
-    const zone = page2.locator(".render-zone");
-    const geom = () =>
-      zone.evaluate((el) => ({ top: el.scrollTop, h: el.scrollHeight, view: el.clientHeight }));
-    const sendPlan = async () => {
-      const before = await page2.locator(".turn-user", { hasText: "plan it step by step" }).count();
-      await page2.locator("textarea").fill("plan it step by step");
-      await page2.keyboard.press("Enter");
-      await page2.waitForFunction(
-        (n) =>
-          [...document.querySelectorAll(".turn-user")].filter((el) =>
-            el.textContent?.includes("plan it step by step"),
-          ).length > n,
-        before,
-        { timeout: 15_000 },
-      );
-    };
-
-    // Three completed turns supply deterministic scrollback; the next is the
-    // live stream under test. The document view's denser paragraph rhythm
-    // leaves two turns only 277px taller than this fixed viewport, too little
-    // for a meaningful 100px upward-wheel proof with 200px left below it.
-    for (let i = 0; i < 3; i++) {
-      const completeBefore = await page2
-        .locator(".turn-assistant", { hasText: "Plan complete — all four steps done." })
-        .count();
-      await sendPlan();
-      await page2.waitForFunction(
-        (n) =>
-          [...document.querySelectorAll(".turn-assistant")].filter((el) =>
-            el.textContent?.includes("Plan complete — all four steps done."),
-          ).length > n,
-        completeBefore,
-        { timeout: 30_000 },
-      );
-      await page2.waitForFunction(() => !document.querySelector(".stop-btn"), undefined, {
-        timeout: 15_000,
-      });
-    }
-    const seeded = await geom();
-    assert.ok(
-      seeded.h - seeded.view > 300,
-      `seed turn did not make scrollback (content ${seeded.h}px, viewport ${seeded.view}px)`,
-    );
-
-    const todoBefore = await page2.locator(".rc-todos").count();
-    const finalBefore = await page2
-      .locator(".turn-assistant", { hasText: "Plan complete — all four steps done." })
-      .count();
-    await sendPlan();
-    await page2.waitForFunction(
-      (n) => document.querySelectorAll(".rc-todos").length > n,
-      todoBefore,
-      { timeout: 15_000 },
-    );
-
-    // A real wheel up over the transcript: the reader is steering into
-    // scrollback while the agent is still producing output.
-    await zone.hover();
-    const atWheel = await geom();
-    await page2.mouse.wheel(0, -1_000);
-    await page2.waitForFunction(
-      ({ top }) => {
-        const el = document.querySelector(".render-zone");
-        return Boolean(
-          el &&
-            el.scrollTop < top - 100 &&
-            el.scrollHeight - el.scrollTop - el.clientHeight > 200,
-        );
-      },
-      atWheel,
-      { timeout: 5_000 },
-    );
-    const before = await geom();
-
-    // Clicking back into the session focuses the prompt without moving the
-    // reader even one pixel or re-arming follow-tail. The subsequent growth
-    // assertion proves output still lands below the detached viewport.
-    await page2.locator("textarea").evaluate((element) => {
-      const textarea = element as HTMLTextAreaElement;
-      const nativeFocus = textarea.focus.bind(textarea);
-      textarea.focus = (options?: FocusOptions) => {
-        textarea.dataset.focusPreventScroll = String(options?.preventScroll === true);
-        nativeFocus(options);
-      };
-    });
-    await zone.click({ position: { x: 2, y: 2 } });
-    assert.equal(
-      await page2.evaluate(() => document.activeElement?.matches(".prompt-box textarea")),
-      true,
-    );
-    assert.equal(
-      await page2.locator("textarea").getAttribute("data-focus-prevent-scroll"),
-      "true",
-      "transcript focus did not request the browser's no-scroll focus mode",
-    );
-    assert.equal((await geom()).top, before.top, "focusing the prompt moved transcript scrollTop");
-
-    // New output must grow below without moving the scrolled-up reader.
-    await page2.waitForFunction(
-      ({ top, h }) => {
-        const el = document.querySelector(".render-zone");
-        return Boolean(el && el.scrollHeight > h && Math.abs(el.scrollTop - top) <= 1);
-      },
-      before,
-      { timeout: 5_000 },
-    );
-
-    // Let that timed turn finish, then use a permission request as a latch
-    // for the re-arm half. Sending the request legitimately returns to the
-    // tail; scroll up once more while the engine is guaranteed to remain
-    // busy until this test answers.
-    await page2.waitForFunction(
-      (n) =>
-        [...document.querySelectorAll(".turn-assistant")].filter((el) =>
-          el.textContent?.includes("Plan complete — all four steps done."),
-        ).length > n,
-      finalBefore,
-      { timeout: 30_000 },
-    );
-    await page2.waitForFunction(() => !document.querySelector(".stop-btn"), undefined, {
-      timeout: 15_000,
-    });
-    await page2.locator("textarea").fill("run something dangerous");
-    await page2.keyboard.press("Enter");
-    await page2.waitForSelector(".perm-bar", { timeout: 15_000 });
-
-    await zone.hover();
-    const latchedAtWheel = await geom();
-    await page2.mouse.wheel(0, -1_000);
-    await page2.waitForFunction(
-      ({ top }) => {
-        const el = document.querySelector(".render-zone");
-        return Boolean(
-          el &&
-            el.scrollTop < top - 100 &&
-            el.scrollHeight - el.scrollTop - el.clientHeight > 200,
-        );
-      },
-      latchedAtWheel,
-      { timeout: 5_000 },
-    );
-
-    // A real downward gesture that reaches the current tail arms from its
-    // pre-input geometry. The unanswered permission keeps the premise
-    // stable; no timer can end the turn between the gesture and assertion.
-    const beforeReturn = await geom();
-    await page2.mouse.wheel(0, beforeReturn.h + beforeReturn.view);
-    await page2.waitForFunction(
-      () => {
-        const el = document.querySelector(".render-zone");
-        return Boolean(
-          el &&
-            document.querySelector(".stop-btn") &&
-            document.querySelector(".perm-bar") &&
-            el.scrollHeight - el.scrollTop - el.clientHeight <= 60,
-        );
-      },
-      undefined,
-      { timeout: 5_000 },
-    );
-    const rearmed = await geom();
-
-    // Allowing resolves the latch and starts tool/output frames without a
-    // new user prompt (which would arm following independently). Do not
-    // merely prove one scroll landed: that later content must carry the
-    // viewport with it.
-    await page2.locator(".perm-allow").click();
-    await page2.waitForFunction(
-      (h) => {
-        const el = document.querySelector(".render-zone");
-        return Boolean(
-          el &&
-            el.scrollHeight > h &&
-            el.scrollHeight - el.scrollTop - el.clientHeight <= 60,
-        );
-      },
-      rearmed.h,
-      { timeout: 5_000 },
-    );
-    await page2.waitForFunction(() => !document.querySelector(".stop-btn"), undefined, {
-      timeout: 15_000,
-    });
-    await page2.waitForFunction(() => {
-      const el = document.querySelector(".render-zone");
-      return Boolean(el && el.scrollHeight - el.scrollTop - el.clientHeight <= 60);
-    });
-  });
-});
-
-test("an overflowing prose transcript supports keyboard scrolling and End re-arms following", async () => {
-  await withFreshMockSession("e2e-transcript-keyboard-9c2f", async (page2) => {
-    await page2.setViewportSize({ width: 900, height: 520 });
-    await page2.waitForSelector("textarea");
-    const zone = page2.locator(".render-zone");
-    const geom = () =>
-      zone.evaluate((el) => ({ top: el.scrollTop, h: el.scrollHeight, view: el.clientHeight }));
-
-    // Make deterministic overflow out of inert response prose. Submitted
-    // inputs now intentionally carry the Phase-IH arrow pairs; beyond those,
-    // the scroller itself remains the keyboard PageUp/End access path.
-    for (let i = 0; i < 8; i++) {
-      const before = await page2.locator(".notice-line[data-source]").count();
-      await page2.locator("textarea").fill(`notice attribution ${i}`);
-      await page2.keyboard.press("Enter");
-      await page2.waitForFunction(
-        (n) => document.querySelectorAll(".notice-line[data-source]").length > n,
-        before,
-        { timeout: 15_000 },
-      );
-      await page2.waitForFunction(() => !document.querySelector(".stop-btn"), undefined, {
-        timeout: 15_000,
-      });
-    }
-    const seeded = await geom();
-    assert.ok(
-      seeded.h - seeded.view > 300,
-      `notice turns did not make scrollback (content ${seeded.h}px, viewport ${seeded.view}px)`,
-    );
-    assert.equal(await zone.getAttribute("tabindex"), "0");
-    const descendantControls = await zone
-      .locator("a, button, input, select, textarea, [tabindex]")
-      .count();
-    const navigationControls = await zone.locator(".input-nav-arrow").count();
-    assert.equal(navigationControls, 16, "each of eight submitted inputs needs both arrows");
-    assert.equal(
-      descendantControls,
-      navigationControls,
-      "fixture gained a transcript control outside submitted-input navigation",
-    );
-    await assertAxeClean(page2, "overflowing prose-only transcript");
-
-    // Hold a turn open so output after End can prove follow-tail was re-armed,
-    // not just that the browser performed one isolated scroll.
-    await page2.locator("textarea").fill("run something dangerous");
-    await page2.keyboard.press("Enter");
-    await page2.waitForSelector(".perm-bar", { timeout: 15_000 });
-
-    await zone.focus();
-    const focus = await zone.evaluate((el) => {
-      const style = getComputedStyle(el);
-      return {
-        active: document.activeElement === el,
-        visible: el.matches(":focus-visible"),
-        outline: style.outlineStyle,
-      };
-    });
-    assert.ok(focus.active, "transcript did not accept focus");
-    assert.ok(focus.visible, "transcript focus ring was not keyboard-visible");
-    assert.notEqual(focus.outline, "none", "transcript has no visible focus outline");
-
-    const atBottom = await geom();
-    await page2.keyboard.press("PageUp");
-    await page2.waitForFunction(
-      ({ top }) => {
-        const el = document.querySelector(".render-zone");
-        return Boolean(
-          el &&
-            el.scrollTop < top - 100 &&
-            el.scrollHeight - el.scrollTop - el.clientHeight > 200,
-        );
-      },
-      atBottom,
-      { timeout: 5_000 },
-    );
-
-    await page2.keyboard.press("End");
-    await page2.waitForFunction(
-      () => {
-        const el = document.querySelector(".render-zone");
-        return Boolean(el && el.scrollHeight - el.scrollTop - el.clientHeight <= 60);
-      },
-      undefined,
-      { timeout: 5_000 },
-    );
-    const rearmed = await geom();
-
-    await page2.locator(".perm-allow").click();
-    await page2.waitForFunction(
-      (h) => {
-        const el = document.querySelector(".render-zone");
-        return Boolean(
-          el &&
-            el.scrollHeight > h &&
-            el.scrollHeight - el.scrollTop - el.clientHeight <= 60,
-        );
-      },
-      rearmed.h,
-      { timeout: 5_000 },
-    );
-    await page2.waitForFunction(() => !document.querySelector(".stop-btn"), undefined, {
-      timeout: 15_000,
-    });
-  });
-});
-
 // The frame sampler's in-page globals (armed before the prompt, read after).
-type BusyWatch = {
-  __busyFrames: number;
-  __blankFrames: number;
-  __glyphFlips: number;
-  __lastGlyph: string;
-  __watch: number;
-};
 
 /** These tests measure a turn from its START, so they must not open while a
  *  previous test's turn is still in flight — an already-ending turn yields a
@@ -2884,89 +1457,6 @@ const awaitIdle = () =>
   page.waitForFunction(() => !document.querySelector(".stop-btn"), undefined, {
     timeout: 30_000,
   });
-
-test("a busy turn never looks idle: the indicator is up, moving, and on screen whenever the stop button is", async () => {
-  // Own the state under test. The old shared-page/checklist version waited
-  // for a transient line, then made another Playwright round-trip to read it;
-  // a loaded runner could deliver the scripted turn_end between those calls,
-  // correctly remove the line, and make textContent() wait 30 seconds for an
-  // element that was supposed to stay gone. A permission request is the
-  // mock's deterministic latch: the turn cannot end until this test answers.
-  await withFreshMockSession("e2e-busy-indicator-9c2f", async (page2) => {
-    await page2.waitForSelector("textarea");
-
-    // Frame-by-frame watcher, armed BEFORE the prompt goes out: any frame
-    // where the turn is in flight (stop button present) but no activity line
-    // is painted is the 2026-07-28 bug — work happening with nothing showing.
-    // It also counts glyph frame CHANGES: a present-but-frozen line is the
-    // 2026-07-29 bug. The callback stays anonymous because tsx's keepNames
-    // wrapper does not exist inside the page.
-    await page2.evaluate(() => {
-      const w = window as unknown as BusyWatch;
-      w.__busyFrames = 0;
-      w.__blankFrames = 0;
-      w.__glyphFlips = 0;
-      w.__lastGlyph = "";
-      w.__watch = window.setInterval(() => {
-        if (document.querySelector(".stop-btn")) {
-          w.__busyFrames++;
-          const glyph = document.querySelector(".activity-glyph")?.textContent;
-          if (glyph === undefined) w.__blankFrames++;
-          else if (glyph !== w.__lastGlyph) {
-            if (w.__lastGlyph !== "") w.__glyphFlips++;
-            w.__lastGlyph = glyph;
-          }
-        }
-      }, 16);
-    });
-
-    await page2.locator("textarea").fill("run something dangerous");
-    await page2.keyboard.press("Enter");
-    await page2.waitForSelector(".perm-bar", { timeout: 15_000 });
-    // The permission bar holds the busy state open, so this waits for real
-    // movement rather than betting that a timed mock reply stays alive long
-    // enough for two browser round-trips.
-    await page2.waitForFunction(
-      () => (window as unknown as BusyWatch).__glyphFlips >= 3,
-      undefined,
-      { timeout: 5_000 },
-    );
-
-    // Scroll and snapshot in one page task while the latch is still held.
-    // The elapsed text and viewport placement cannot disappear between two
-    // Playwright calls now.
-    const held = await page2.evaluate(() => {
-      document.querySelector(".render-zone")!.scrollTop = 0;
-      const line = document.querySelector(".activity-line");
-      const r = line?.getBoundingClientRect();
-      return {
-        text: line?.textContent ?? "",
-        visible: Boolean(r && r.height > 0 && r.top >= 0 && r.bottom <= window.innerHeight),
-      };
-    });
-    assert.match(held.text, /\(\d+s\)/, "no elapsed counter in the activity line");
-    assert.ok(held.visible, "indicator off screen while the transcript is scrolled up");
-
-    await page2.locator(".perm-deny").click();
-    await page2.waitForFunction(() => !document.querySelector(".stop-btn"), undefined, {
-      timeout: 15_000,
-    });
-    await page2.waitForSelector(".activity-line", { state: "detached", timeout: 5_000 });
-    const frames = await page2.evaluate(() => {
-      const w = window as unknown as BusyWatch;
-      window.clearInterval(w.__watch);
-      return { busy: w.__busyFrames, blank: w.__blankFrames, flips: w.__glyphFlips };
-    });
-    assert.ok(frames.busy > 0, "the sampler never saw the turn in flight");
-    assert.equal(
-      frames.blank,
-      0,
-      `${frames.blank} frame(s) had a turn in flight with no activity line painted`,
-    );
-    assert.ok(frames.flips >= 3, `the glyph moved only ${frames.flips} time(s) while held busy`);
-    assert.equal(await page2.locator(".activity-line").count(), 0);
-  });
-});
 
 // In-page globals for the queued-follow-up sampler below.
 type QueueWatch = { __framesSeen: boolean[]; __qWatch: number };
@@ -3247,9 +1737,19 @@ test("E.5: expanded dirs survive a close/reopen, and a turn's auto-refresh keeps
   await page.keyboard.type("plan it step by step");
   await page.keyboard.press("Enter");
   await page.waitForSelector("text=Plan complete — all four steps done.", { timeout: 30_000 });
-  await page.waitForTimeout(400); // let the turn_end refresh land…
-  // …then let it FINISH: the refetch swaps the rows, so a slow one under load
-  // must not read as a collapse.
+  // The turn_end refresh refetches expanded dirs; wait for that fan-out to
+  // stop (no new fs_ frame for a full poll interval) instead of a fixed nap.
+  await page.waitForFunction(
+    () => {
+      const w = window as unknown as { __fsSent?: unknown[]; __fsStable?: number };
+      const n = w.__fsSent?.length ?? 0;
+      const stable = w.__fsStable === n;
+      w.__fsStable = n;
+      return stable;
+    },
+    undefined,
+    { polling: 150, timeout: 10_000 },
+  );
   await eventually(() => !!document.querySelector(".files-panel"), "auto-refresh closed the panel");
   await eventually(
     () =>
@@ -3271,7 +1771,18 @@ test("E2.2: the tree is LAZY — open fetches root + first level only; expand fe
   // request is depth ≤ 1, and the whole-tree fs_list is never sent.
   await page.locator(".ab-files").click();
   await page.waitForSelector(".files-file-row:has-text('package.json')");
-  await page.waitForTimeout(300); // let the prefetch fan-out finish
+  // The prefetch fan-out is done when no new fs_ frame lands for a poll.
+  await page.waitForFunction(
+    () => {
+      const w = window as unknown as { __fsSent?: unknown[]; __fsStable?: number };
+      const n = w.__fsSent?.length ?? 0;
+      const stable = w.__fsStable === n;
+      w.__fsStable = n;
+      return stable;
+    },
+    undefined,
+    { polling: 150, timeout: 10_000 },
+  );
   const onOpen = await sent();
   assert.ok(onOpen.every((m) => m.type !== "fs_list"), "the whole-tree request is retired from the client");
   const listdirs = onOpen.filter((m) => m.type === "fs_listdir");
@@ -3368,11 +1879,8 @@ test("E2.4: the Projects-root proof — lazy expands into two repos with per-rep
 
   // Seed the session AT the fixture over the wire (the UI has no cwd
   // picker), then join it from the browser like any viewport would.
-  const seed = new TestClient(d.port, { token: TOKEN });
-  await seed.opened();
-  await seed.type("agents");
-  seed.send({ type: "create", agent: "claude-code", cwd: mr } as ClientMsg);
-  const created = (await seed.type("session_created")) as { sessionId: string } & Record<string, unknown>;
+  const { client: seed, sessionId: seededId } = await createSession(d.port, "claude-code", { cwd: mr, token: TOKEN });
+  const created = { sessionId: seededId };
   const backUrl = page.url(); // later tests continue the original session
   try {
     await page.goto(`${base}/s/${created.sessionId}`);
@@ -3494,10 +2002,7 @@ test("W.2: the live tree — a write behind the UI's back appears with zero clic
   mkdirSync(path.join(ws, "colly", "sub"), { recursive: true });
   writeFileSync(path.join(ws, "colly", "sub", "inner.txt"), "inner\n");
 
-  const seed = new TestClient(d.port, { token: TOKEN });
-  await seed.opened();
-  await seed.type("agents");
-  seed.send({ type: "create", agent: "claude-code", cwd: ws } as ClientMsg);
+  const { client: seed, sessionId: liveSeedId } = await createSession(d.port, "claude-code", { cwd: ws, token: TOKEN });
   const created = (await seed.type("session_created")) as { sessionId: string } & Record<string, unknown>;
   const backUrl = page.url();
   try {
@@ -3561,44 +2066,6 @@ test("W.2: the live tree — a write behind the UI's back appears with zero clic
     await page.waitForSelector("textarea");
     rmSync(ws, { recursive: true, force: true });
   }
-});
-
-test("a notice in the engine's own words is badged; the shell's own words aren't", async () => {
-  // Own the session whose output is under test. The shared page has just
-  // navigated through two Explorer fixtures; under runner load its attach /
-  // replay can still be settling when this prompt is typed, so a timeout can
-  // happen before the notice path runs and say nothing about attribution.
-  await withFreshMockSession("e2e-notice-attribution-9c2f", async (page2) => {
-    await page2.waitForSelector("textarea");
-    await page2.locator("textarea").fill("show me a notice");
-    await page2.keyboard.press("Enter");
-    await page2.waitForSelector(".notice-line[data-source]", { timeout: 15_000 });
-
-    // The engine's line carries its name and no shell glyph…
-    const engine = page2.locator(".notice-line[data-source]").last();
-    assert.equal(await engine.getAttribute("data-source"), "mock-engine");
-    assert.equal(await engine.locator(".notice-source").innerText(), "mock-engine");
-    assert.equal(await engine.locator(".notice-glyph").count(), 0);
-    assert.match(await engine.innerText(), /re-enter your API key/);
-
-    // …and Mirafold's own line carries the glyph and no badge, so the two can't
-    // be confused: an engine string can't render as the shell speaking (2026-07-20).
-    const shell = page2
-      .locator(".notice-line:not([data-source])")
-      .filter({ hasText: "context compacted" })
-      .last();
-    assert.equal(await shell.locator(".notice-source").count(), 0);
-    assert.equal(await shell.locator(".notice-glyph").count(), 1);
-    // The difference is visible, not just structural.
-    assert.equal(
-      await engine.evaluate((el) => getComputedStyle(el).borderLeftStyle),
-      "dashed",
-    );
-    assert.equal(
-      await shell.evaluate((el) => getComputedStyle(el).borderLeftStyle),
-      "solid",
-    );
-  });
 });
 
 // C.2 — the automated Phase-A regression guard; assertAxeClean (scope, tags,
